requirements.txt
aioredis==1.3.1
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
Django==2.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
flake8==3.8.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
mccabe==0.6.1
msgpack==1.0.0
mysqlclient==2.0.1
Pillow==8.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dotenv==0.14.0
pytz==2020.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
zipp==3.3.2
zope.interface==5.1.2
